+++ src/starbase/exceptions.py
@@ -2,7 +2,10 @@ __title__ = 'starbase.exceptions'
 __author__ = 'Artur Barseghyan'
 __copyright__ = 'Copyright (c) 2013 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
-__all__ = ('ImproperlyConfigured', 'InvalidArguments', 'ParseError', 'DoesNotExist')
+__all__ = (
+    'BaseException', 'ImproperlyConfigured', 'InvalidArguments', 'ParseError', 'DoesNotExist',
+    'DatabaseError', 'IntegrityError'
+)
 
 class BaseException(Exception):
     """
@@ -28,7 +31,19 @@ class ParseError(BaseException):
     """
 
 
-class DoesNotExist(BaseException):
+class DatabaseError(BaseException):
+    """
+    General database error, as defined in PEP249 interface.
+    """
+
+
+class DoesNotExist(DatabaseError):
     """
     Does not exist.
+    """
+
+
+class IntegrityError(DatabaseError):
+    """
+    Integrity error, as defined in PEP249 interface.
     """